requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
requests
httpx
pydantic
fitz